Added toolchain_msvc.cmake
This is the first step of porting the compilation to windows.

Warning level is set to a high level (/W4), and some additional warning were enabled.

As of now, one warning needs to be disabled, concerning variable shadiowing
  4459 # declaration of 'xs' hides global declaration
(since the xs variable name is quite commonly used)

Also : gitignore build64