conanfile.py : can install doctest from conan
There is a new cmake option (FPLUS_UNITTEST_USE_CONAN) that enables
to install doctest via conan.

Usage:
> cd build/
> conan install .. -obuild_unittest=True
> cmake .. -DFPLUS_BUILD_UNITTEST=ON -DFPLUS_UNITTEST_USE_CONAN=ON

By default, the conanfile will *not* intall doctest
(an option 'build_unitest' is required).

Note:
I also had to add two libraries for the link (-lstdc++ and -lm)
May be this is only required for osx (but it does not break the linux build)